Remove all packages that depend on packages from the Emacswiki
These commits were removed in the parent commit (which see).

Only remove packages that hard `require' a package from the Emacswiki
at the top-level and that specify the dependency using the
Package-Requires header keyword.  There are some other packages that
soft require such a package and/or only require it when some command
is called and/or when some non-essential library is loaded, and that
do not specify a dependency using Package-Requires.

Also see https://github.com/melpa/melpa/pull/5008.